@@ -22,24 +22,24 @@
   },
   "homepage": "https://devne.ws/",
   "dependencies": {
-    "autolinker": "^1.4.0",
+    "autolinker": "^1.4.4",
+    "babel-preset-react-hmre": "^1.1.1",
     "html-escape": "^2.0.0",
-    "localforage": "^1.4.2",
-    "material-ui": "^0.16.4",
-    "moment": "^2.12.0",
-    "normalize.css": "^5.0.0",
-    "react": "^15.0.1",
-    "react-back2top": "^0.1.2",
-    "react-dom": "^15.0.1",
-    "react-tap-event-plugin": "^2.0.1",
-    "superagent": "^3.1.0"
+    "localforage": "^1.5.2",
+    "material-ui": "^0.19.4",
+    "moment": "^2.19.1",
+    "normalize.css": "^7.0.0",
+    "prop-types": "^15.6.0",
+    "react": "^16.0.0",
+    "react-dom": "^16.0.0",
+    "react-tap-event-plugin": "^3.0.2",
+    "superagent": "^3.6.3"
   },
   "devDependencies": {
     "babel-core": "^6.7.2",
     "babel-loader": "^6.2.4",
-    "babel-preset-es2015": "^6.6.0",
+    "babel-preset-env": "^1.6.0",
     "babel-preset-react": "^6.5.0",
-    "babel-preset-react-hmre": "^1.1.1",
     "copy-webpack-plugin": "^4.0.1",
     "css-loader": "^0.26.0",
     "html-webpack-plugin": "^2.12.0",
@@ -50,7 +50,7 @@
     "postcss-import": "^8.0.2",
     "postcss-loader": "^1.1.1",
     "raw-loader": "^0.5.1",
-    "react-svg-inline": "^1.1.0",
+    "react-svg-inline": "^2.0.1",
     "style-loader": "^0.13.0",
     "surge": "^0.18.0",
     "webpack": "^1.12.14",
